print certficate collector config in v2 to achieve v1 behaviour
